Revert "[TASK] Make clearable button more accessible"
This reverts commit 2a05d2844d514918ec2c73cdeab9fcab199d222e.

The proposed and merged solution decreases the overall usability in the
TYPO3 backend as users are now forced to press the <TAB> key twice to
actually switch the focus to another field.

One could also argue that the change that is reverted is not suitable
for the target group that mostly uses keyboard navigation as said users
were already able to clear fields previously.

Once we have a solution that satisfies both target groups, another
patch will be pushed.

Change-Id: I58acdd89a61f3df258e755c893389fe960985b8f
Resolves: #102504
Reverts: #102372
Releases: main, 12.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81904
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>