[!!!][FEATURE] Add PSR-14 event for manipulating ContentObjectRenderer
This introduces a new PSR-14 event, enabling
extension authors to modify the ContentObjectRenderer
instance, after it has been initialized.

This replaces the previously available hook
$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['tslib/class.tslib_content.php']['postInit']
which is therefore now removed.

The hook required implementations to implement
the `ContentObjectPostInitHookInterface`, which
is now deprecated as it's unused due to the
hook removal.

Resolves: #102581
Releases: main
Change-Id: I86501b61029d69e05703593af3903ae9d7851ea0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82063
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>