[!!!][FEATURE] Add PSR-14 event for modifying image source collection
This introduces a new PSR-14 event, enabling
extension authors to modify the image source
collection of an ImageContentObject.

This replaces the previously available hook
$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['tslib/class.tslib_content.php']['getImageSourceCollection']
which is therefore now removed.

The hook required implementations to implement
the `ContentObjectOneSourceCollectionHookInterface`,
which is now deprecated as it's unused due to the
hook removal.

Resolves: #102624
Releases: main
Change-Id: Id3c26a613984a673a118f4fc5acbc3528e3a5f5a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82145
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>