[TASK] Adapt to testing-framework not loading EXT:install
typo3/testing-framework no longer loads EXT:install
in functional tests by default. The patch adapts a
couple of tests to cope with this.

Note core will continue to refactor in this area
after 13.0: It for instance needs to be sorted out
if the upgrade wizard infrastructure should be
moved to EXT:core. For this reason, no full
communication happened in this area, yet.

> composer u typo3/testing-framework

Resolves: #102982
Releses: main
Change-Id: Id786a2f16d336b6e9d80519e7f944d560f61d8bd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82695
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>