[BUGFIX] Install Tool: Load renderable components in correct scope
Lit elements need to get created in the correct browser scope
(i.e. frame) to get rendered in Firefox, which is a huge issue in
modals (that reside in the global scope), that are triggered in an
iframe scope.

This commit introduces a helper method `loadModuleFrameAgnostic()` that
loads the module depending on the context the Install Tool is in
(standalone vs. iframe).

All Install Tool modules are modified to load renderable modules (info
boxes, flash messages and progress bars) explicitly. Also, a superfluous
progress bar is removed from the "Directory status" module that breaks
the whole modal in case a user prefers reduced motions.

Resolves: #103048
Releases: main
Change-Id: I00834e1000c689d4581cbd3ef0124a1f8a391186
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82791
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Benjamin Franzke <ben@bnf.dev>