[TASK] Show sender email address for mailtest in ext:install
In TYPO3 versions before v9, the mailtest showed the configured
sender email address. This has been removed due to a wrong fluid
variable usage with #88523

With this change, the sender email address is now shown again in
the mailtest. Additionally, a note is added to clarify, that not
receiving the test email could be due to spam filters or the
sender domain SPF settings.

The notice about the `defaultMailFromAddress` has been removed,
since it is already checked in `EnvironmentController`, that
the value is configured and a valid email address.

Finally, the fluid variable `queueIdentifier` introduced with
#100084 has been removed, since it is not used in the template.

Resolves: #103136
Related: #88523
Related: #100084
Releases: main, 12.4
Change-Id: Ib076934732a05d9c709f6d822526e8316cf87106
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83021
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>