[TASK] Clean up legacy cache_treelist leftovers
Using cache_treelist has been avoided with #103139
since it was broken. v13 now goes ahead and removes
the clearing logic and the table itself.

Note the PageRepository method that formerlly used
that cache would be a great target for a (first) CTE
to structurally enhance performance. v13 may see
patches in this area.

Resolves: #103165
Related: #103139
Releases: main
Change-Id: I2bcee1befa08b8dc81f39c7ef4b9e8a005fad0c8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83051
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>