[BUGFIX] Create real line break elements in Upgrade Wizard view
The "Upgrade Wiards" view automatically checks whether mandatory
database changes are pending. In that case, a list of tables, fields and
indices are rendered. Before the removal of jQuery, a line break was
rendered between the list items. However, the migration was not
complete, as the line breaks are now rendered as `<br>`.

The commit fixes this by creating real `<br>` elements.

Resolves: #103613
Releases: main
Change-Id: I6896ceeaed30606f55ac4376002552f92224c878
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83784
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>