[TASK] Provide list of available locales in Install Tool
The Site Configuration module currently allows to
pick a language locale from a list of available locales
of the system (via the "locale -a" command).

This patch moves this functionality into a reusable
method, which is now also utilized from within
the Install Tool when configuring the
`TYPO3_CONF_VARS[SYS][locale]` option. The description
of this "locale" option now holds all possible
locale variants.

If the "locale -a" execution fails, the output will
show an empty list.

Resolves: #61189
Releases: main, 12.4
Change-Id: Icf048be100e8e658fb14bb983d7bab5101bd1587
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85533
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>