[BUGFIX] Fix ExtensionScanner ignoring all lines instead of single one
Once the ExtensionScanner hits a '@extensionScannerIgnoreLine'
annotated code line, it internally sets a flag that is never
reset.

That causes all following lines to be ignored, not just one.

This patch will reset the property `$isCurrentLineIgnored`
to `false` again, once the annotated line has passed.

Tests are adapted to verify this behaviour.

Due to the behavioral change it is possible that the scanner
now finds more hits in third-party scanned extensions
that previously may have been ignored.

Resolves: #100702
Releases: main, 12.4
Change-Id: Ib1cbb5e7758e7e302e4395d759e8585eeff15d8e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84733
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>