[BUGFIX] Make BackendGroupsExplicitAllowDeny wizard indicate errors
The migration wizard BackendGroupsExplicitAllowDeny
('Migrate backend groups "explicit_allowdeny" field to
simplified format') returns Symfony formatted error messages
for specific single rows.

Before this patch, it was still reported as a successful
migration, and the success message is formatted with
"everything is great" green background, so that errors
may not receive attention.

This patch changes the wizard to return a failure state.
This will then return the messages formatted as errors.

Since the wizard migrates all rows (even the faulty ones),
the wizard can then be re-executed to be marked as "done".

This is not ideal, but because of the way Symfony
output streams cannot reveal if they contain an error
when the stream is retrieved, there is little other choice.

Otherwise the Wizard API would need to receive a breaking
API change to allow to carry additional meta information
about "warnings" rather than just errors, which is
beyond scope for current development state.

Resolves: #104208
Related: #104225
Releases: main, 12.4
Change-Id: Ibde1d977bc8747a053da1360aea6c0e3f8b77f7f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84933
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>