[BUGFIX] Fix broken Installer when typo3conf dir is missing
With issue #104074 a change was introduced that no longer required
"typo3conf" to exist in composer mode.

Sadly the installer performed a write check on the specific
old "typo3conf" location instead of using the actual
new location for config files since TYPO3 v12.

The installer does not report back FlashMessages on failed
AJAX requests, also those AJAX requests currently oddly
do not even show up in browser's network panel. Without
this, an enduser is not able to finish the installation,
which could be done by just creating an empty
"typo3conf" directory. This should be addressed in a
follow-up, the FlashMessage is already prepared now
to be contained in the JSON response.

Resolves: #104306
Related: #104074
Releases: main
Change-Id: Icabdee13e79f016ef17cdd3e2ab62cf078f63b4c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85106
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>