[BUGFIX] Create default backend user groups without EXT:dashboard
This patch corrects the optional last step of the initial setup
process "Create default backend user groups" if EXT:dashboard is
not installed. This may be the case for a composer based
installation. Instead of silently failing the step due to the
missing column "be_groups.availableWidgets" this permission is
not written to the database.

Resolves: #104434
Releases: main
Change-Id: I20216cb9032a36ad36a862289ba98082249b0a26
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85300
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>