[BUGFIX] Do not enforce color scheme in login screen and Install Tool if not set
As the color scheme feature, especially dark mode is still experimental,
the `PageRenderer` now checks whether the backend user is authenticated
before applying the data attribute to control the rendering.

The same logic is now applied in the Install Tool running in the backend
context, as it wasn't checked whether the feature is enabled before.

Resolves: #104985
Releases: main
Change-Id: I0b988a0a9a9f3a6a765b1f883aaecc96eeb2626a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86118
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>