[TASK] Make TYPO3 logo in dashboard, about and installer context aware
The SVG logo is now styled based on the inherited color scheme
feature [1], that allows embedded documents to use
`@media (prefers-color-scheme: dark)` to detect the value of the
`color-scheme` attribute of the embedding document.
That means it takes the users color-scheme setting (auto, light or dark)
into account.

The previous typo3_orange.svg is kept as-is for the SystemEmail which
is not color-scheme aware.

[1] https://developer.mozilla.org/en-US/docs/Web/CSS/@media/prefers-color-scheme#inherited_color_scheme_in_embedded_elements

Resolves: #105138
Releases: main
Change-Id: I8dbb6eb298cded42b2cbd70b060feed4b710fc94
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86354
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>