[BUGFIX] Setup: Normalize "no" as input when asking for basic site
The `setup` command asks whether a basic site shall be created when
installing TYPO3. The question implies that it defaults to "no" value,
but fails to accept "no" as input given by the user.

This commit adds a normalizer for that question that converts "no" to an
empty string, having the same effect as stepping through the question by
providing no input at all.

Resolves: #105203
Releases: main, 12.4
Change-Id: Ifcbccfa0c3c17304412112318ffc5bb0f0e87790
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86446
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Garvin Hicking <gh@faktor-e.de>