[BUGFIX] Properly scan for breaking GU::sanitizeLocalUrl() change
With #101311 the function argument for
`GeneralUtility::sanitizeLocalUrl` has been declared as required.
The breaking change is however not detected by extension scanner,
because the associated extension scanner configuration has been
added to `MethodArgumentRequiredMatcher.php`. Since the changed
function is static, the correct file for the extension scanner
configuration is `MethodArgumentRequiredStaticMatcher.php`.

This change moves the associated extension scanner configuration
to the correct file.

Resolves: #105335
Related: #101311
Releases: main
Change-Id: I24d3209cb6822ddfdb3a1611e5dfd66d2147dd99
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86640
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>