[TASK] Add new list_type deprecation match to extension scanner
With #105076, the plugin content element and plugin sub types
have been deprecated. The extension scanner currently only
matches the deprecation, when it detects the usage of
`ExtensionUtility::PLUGIN_TYPE_PLUGIN` as fifth argument
for `ExtensionUtility::configurePlugin()`. Many public
extensions however omit the fifth argument, resulting
in the deprecation not being shown in extension scanner
and extension developers not being aware of the required
code changes.

This change adds a new entry to the
`MethodArgumentRequiredStaticMatcher.php` configuration, so
the extension scanner will show a match for the deprecation,
if `ExtensionUtility::configurePlugin()` is called without
a fifth argument.

Resolves: #105480
Related: #105076
Releases: main, 13.4
Change-Id: Ic5b7517b9eb072bcbdb1ec91a59c519ae6a9dfef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86797
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>