[BUGFIX] Adapt migration for changed folder_identifier SQL definition
With the introduction of automatic SQL column generation for type
`folder`, the migration for table sys_file_collection and field
folder_identifier broke due to a changed default value. Before the
change the default value was an empty string. Now it is NULL.

The new field `folder_identifier` got added with #97923 plus
the migration. This means users who are upgrading from v11 to v13
and skipping v12 will have the new default `NULL`.

The query is adapted to look for both the old and new default value.

Resolves: #105601
Related: #101890
Related: #97923
Releases: 13.4
Change-Id: Ife0d3ba321b5c14c72a5928c2c2c5309d6e141f0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86993
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>