[FEATURE] Configure default image conversion/preview formats
A new config option:

  `$GLOBALS['TYPO3_CONF_VARS']['GFX']['imageFileConversionFormats']`

is added that allows to define two things:

1) Define a list (key=inputFileExtension => value=outputFileExtension)
   of allowed/supported target conversion formats
   for a TYPO3 instance.

2) Define which file formats get converted to specific output formats
   by default (special array key "default").

This allows forward compatibility to new file formats, and to allow
instance maintainers to use a format like "webp" or "avif" for
all image conversions that do not specify a target fileExtension.

Before, the TYPO3 core internally picked jpeg or png (depending
on use of transparency) without the possibility to influence this
without the use of specific extensions.

Some default file formats are now converted to png instead of jpeg
due to the "default => png" setting. Tests for this are adjusted to
match this new behavior. Preference for png has been made to keep
the option of transparency information. It is recommended for instances
that support writing webp/avif to set default => webp or default => avif
for optimized file sizes.

Based on a patch by Philipp Kitzberger:
https://review.typo3.org/c/Packages/TYPO3.CMS/+/78260

The old `GraphicalFunctions->gif_or_jpg()` method is deprecated
and superseded by
`GraphicalFunctions->determineDefaultProcessingFileExtension()`,
former usages are reported by the extension scanner. The new
replacement method is declared `@internal` and will be
moved to a service class in a follow-up.

Resolves: #93981
Related: #106553
Releases: main
Change-Id: I64c0cfdaa1b7deb067dd2ecc9cd207c36c0c92d1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87582
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>