[TASK] Add ScalarStringMatcher Extension Scanner for TYPO3_MODE reports
An upgrade to TYPO3 v12 can be painful when backend or frontend
modules (or even the Database Analyzer) bails out with a white
page. Usually this would lead to inspection of logfiles,
checking if PHP's `error_reporting` or `display_errors` is
not properly set, looking for PHP/Webserver coredumps,
problems in HTTP headers, CSP errors and much more, just
to find out that a

  `defined('TYPO3_MODE') or die()`

was left in some PHP file of an extension, which needs
to be replaced with `TYPO3` (see #92947 and #96107).

The Extension Scanner only scanned for the full
PHP constant usage before.

With this patch, these "magic strings" can now also
be scanned and reported, making debugging of a
malfunctioning upgrade much easier for developers
and especially integrators.

Resolves: #105898
Related: #92947
Related: #96107
Releases: main, 13.4, 12.4
Change-Id: I1dc9cebee1773e9c40da69831640486b02615911
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87663
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>