[TASK] Test transparent image processing with avif and webp
The image formats avif and webp support transparency. For both
formats new tests are added to the image processing tests in the
install tool. They convert a png with transparency to avif/webp
to confirm the correct output.

As of today ddev v1.24.2 web-images are debian bookworm based and
ship with ImageMagick 6.9.11-60 Q16. This version is known to
produce black backgrounds with avif. ImageMagick 7 renders the
transparency correctly.

Resolves: #106004
Releases: main, 13.4
Change-Id: I0334c1fcc910f85484f880779291507615279cc6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87924
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>