[TASK] Use abstract docs.typo3.org linking service method
The DocumentationViewHelper contains logic to be able to link
to typo3.docs.org with specific versioning information. This
logic was not available to other areas in the PHP scope of
the core.

With this patch, this logic is put into the Typo3Information
DO, where other similar URLs are already listed.

This allows to have one central place to adjust documentation
URLs and the DocumentationViewHelper can now utilize this via:

```
$url = (new Typo3Information())
  ->getDocsLink('changelog:important-106192-1739862406');
```

(or using dependency injection)

As a drive-by, the URL_OPCACHE contents have been adjusted
and usage of existing instances is enforced.
Note that DI is not possible within the ViewHelper because
it is used in Install-Tool context that does not provide full DI.

Resolves: #106373
Releases: main, 13.4
Change-Id: I8a71672a601b7a9d7bf9b3f6e40ab6d2982c59f4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88633
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>