[BUGFIX] Preserve transparency for image/graphicsmagick conversions
By default, imagemagick and graphicsmagick will convert for example
a SVG image with transparency to a PNG image by replacing transparency
information with white (see [1]).

To preserve transparency, the option "-background none" is added
by default to the list of parameters.

Note that this can also be influenced by using the TYPO3 configuration
option `procvessor_stripColorProfileParameters' to include the option
manually:

```
$GLOBALS['TYPO3_CONF_VARS']['GFX']
    ['processor_stripColorProfileParameters'] = [
  '+profile',
  '*',
  '-background',
  'none'
];
```

This is considered a hack, and TYPO3 should provide this
out-of-the-box.

Due to possible side-effects and the ability to fix this with
configuration, this patch only targets main (v14).

Sadly it is not (easily) possible to run automated tests with
different im/gm versions in the testing suite, and to check
transparency information on compressed files is prone to false
negatives.

The "Admin Tools > Environment > Test Images ..." check has been
enhanced with one variant of transparent SVG to PNG and WEBP to
showcase the behaviour (also showcasing the issue that its not
transparent, when the "-background none" option is not added).

For SVG conversion to properly work and retrieve width/height
information for scaling purposes, this patch allows this as
a drive-by. A todo note to optimize this is already in effect
and should be addressed with a dedicated change.

[1] https://imagemagick.org/script/command-line-options.php#background

Resolves: #97975
Releases: main
Change-Id: Ifec146c7680351b0401363415c17c9e416b7efc4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89032
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>