[BUGFIX] Evaluate mixed-case file extensions in mime_type migration
The SysFileMimeTypeMigration wizard evaluates the database
result set of existing mis-matching MIME-type / file extension
pairs so they can be mapped to proper types.

This failed to recognize though that due to the default
collations being set to "ci" (case insensitive) in most
RDBMS setups, the wizard might receive a valid database
result set of a file extensions set to "CSV" (instead of
"csv"). It would then fail to be able to look up the
lowercase-based $mimeTypeCompatibility array.

Resolves: #106805
Related: #106735
Releases: main, 13.4, 12.4
Change-Id: I57a3a3ab639c8de1a3e42d7a04a9aaf1bc9d60bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89677
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>